add composition root and refactor `protostar_cli`
