+++ protostar/start.py
@@ -7,26 +7,26 @@ from protostar.cli import (
     CLIApp,
     MissingRequiredArgumentException,
 )
-from protostar.protostar_cli import ConfigurationProfileCLISchema, ProtostarCLI
+from protostar.composition_root import build_di_container
+from protostar.configuration_profile_cli import ConfigurationProfileCLI
 from protostar.protostar_exception import UNEXPECTED_PROTOSTAR_ERROR_MSG
 
 
 def main(script_root: Path):
-    protostar_cli = ProtostarCLI.create(script_root)
+    di_container = build_di_container(script_root)
+    protostar_cli = di_container.protostar_cli
+    protostar_toml_reader = di_container.protostar_toml_reader
 
     configuration_profile_name = (
-        ArgumentParserFacade(ConfigurationProfileCLISchema(), disable_help=True)
+        ArgumentParserFacade(ConfigurationProfileCLI(), disable_help=True)
         .parse(ignore_unrecognized=True)
         .profile
     )
-
-    parser = ArgumentParserFacade(
-        protostar_cli,
-        default_value_provider=ArgumentValueFromConfigProvider(
-            protostar_cli.protostar_toml_reader,
-            configuration_profile_name=configuration_profile_name,
-        ),
+    argument_value_from_config_provider = ArgumentValueFromConfigProvider(
+        protostar_toml_reader,
+        configuration_profile_name,
     )
+    parser = ArgumentParserFacade(protostar_cli, argument_value_from_config_provider)
 
     try:
         asyncio.run(protostar_cli.run(parser.parse()))
